+++ languagetool-core/src/main/java/org/languagetool/rules/spelling/SpellingCheckRule.java
@@ -24,6 +24,9 @@ import org.languagetool.Language;
 import org.languagetool.rules.ITSIssueType;
 import org.languagetool.rules.Rule;
 import org.languagetool.rules.RuleMatch;
+import org.languagetool.rules.patterns.PatternToken;
+import org.languagetool.rules.patterns.PatternTokenBuilder;
+import org.languagetool.tagging.disambiguation.rules.DisambiguationPatternRule;
 import org.languagetool.tokenizers.WordTokenizer;
 
 import java.io.IOException;
@@ -56,7 +59,8 @@ public abstract class SpellingCheckRule extends Rule {
   private final Set<String> wordsToBeIgnored = new HashSet<>();
   private final Set<String> wordsToBeProhibited = new HashSet<>();
   private final CachingWordListLoader wordListLoader = new CachingWordListLoader();
-
+  
+  private List<DisambiguationPatternRule> antiPatterns = new ArrayList<>();
   private boolean considerIgnoreWords = true;
   private boolean convertsCase = false;
 
@@ -86,13 +90,15 @@ public abstract class SpellingCheckRule extends Rule {
 
   /**
    * Add the given words to the list of words to be ignored during spell check.
+   * You might want to use {@link #acceptPhrases(List)} instead, as only that
+   * can also deal with phrases.
    */
   public void addIgnoreTokens(List<String> tokens) {
     wordsToBeIgnored.addAll(tokens);
   }
 
   /**
-   * Set whether the list of words to be explicitly ignored is considered at all.
+   * Set whether the list of words to be explicitly ignored (set with {@link #addIgnoreTokens(List)}) is considered at all.
    */
   public void setConsiderIgnoreWords(boolean considerIgnoreWords) {
     this.considerIgnoreWords = considerIgnoreWords;
@@ -260,6 +266,33 @@ public abstract class SpellingCheckRule extends Rule {
    */
   protected List<String> expandLine(String line) {
     return Collections.singletonList(line);
+  }
+
+  /**
+   * Accept the given phrases even though they are not in the built-in dictionary.
+   * Use this to avoid false alarms on e.g. names and technical terms. Unlike {@link #addIgnoreTokens(List)}
+   * this can deal with phrases. A way to call this is like this:
+   * <code>rule.acceptPhrases(Arrays.asList("duodenal atresia"))</code>
+   * This way, checking would not create an error for "duodenal atresia", but it would still
+   * create and error for "duodenal" or "atresia" if they appear on their own.
+   * @since 3.3
+   */
+  public void acceptPhrases(List<String> phrases) {
+    List<List<PatternToken>> antiPatterns = new ArrayList<>();
+    for (String phrase : phrases) {
+      String[] parts = phrase.split(" ");
+      List<PatternToken> patternTokens = new ArrayList<>();
+      for (String part : parts) {
+        patternTokens.add(new PatternTokenBuilder().token(part).build());
+      }
+      antiPatterns.add(patternTokens);
+    }
+    this.antiPatterns = makeAntiPatterns(antiPatterns, language);
+  }
+
+  @Override
+  public List<DisambiguationPatternRule> getAntiPatterns() {
+    return antiPatterns;
   }
 
 }