+++ languagetool-language-modules/de/src/test/java/org/languagetool/rules/spelling/hunspell/SpellingCheckRuleTest.java
@@ -21,10 +21,13 @@ package org.languagetool.rules.spelling.hunspell;
 import org.junit.Test;
 import org.languagetool.JLanguageTool;
 import org.languagetool.language.GermanyGerman;
+import org.languagetool.rules.Rule;
 import org.languagetool.rules.RuleMatch;
 import org.languagetool.rules.de.GermanSpellerRule;
+import org.languagetool.rules.spelling.SpellingCheckRule;
 
 import java.io.IOException;
+import java.util.Arrays;
 import java.util.List;
 
 import static org.hamcrest.CoreMatchers.is;
@@ -41,6 +44,21 @@ public class SpellingCheckRuleTest {
     List<RuleMatch> matches = lt.check("Das ist ein Tibbfehla");
     assertThat(matches.size(), is(1));
     assertThat(matches.get(0).getRule().getId(), is(GermanSpellerRule.RULE_ID));
+  }
+
+  @Test
+  public void testIgnorePhrases() throws IOException {
+    JLanguageTool lt = new JLanguageTool(new GermanyGerman());
+    assertThat(lt.check("Ein Test mit Auriensis Fantasiewortus").size(), is(2));
+    for (Rule rule : lt.getAllActiveRules()) {
+      if (rule instanceof SpellingCheckRule) {
+        ((SpellingCheckRule) rule).acceptPhrases(Arrays.asList("Auriensis Fantasiewortus"));
+      } else {
+        lt.disableRule(rule.getId());
+      }
+    }
+    assertThat(lt.check("Ein Test mit Auriensis Fantasiewortus").size(), is(0));
+    assertThat(lt.check("Ein Test mit Auriensis und Fantasiewortus").size(), is(2));  // the words on their own are not ignored
   }
 
 }